Added inventory functionality to userController. Removed Inventory controller and route. Updated itemSchema.
