+++ server/models/store.js
@@ -4,6 +4,9 @@ const Schema = mongoose.Schema;
 // Define the schema for store items
 const StoreItemSchema = new Schema({
     name: String,
+    description: String,
+    category: String,
+    price: Number,
     enabled: {
         type: Boolean,
         default: false // By default, items are not enabled
@@ -15,21 +18,23 @@ const StoreItem = mongoose.model('StoreItem', StoreItemSchema);
 
 // Initialize an array of store items
 const initialStoreItems = [
-    { name: 'Item A', enabled: false },
-    { name: 'Item B', enabled: false },
-    { name: 'Item C', enabled: false },
-    { name: 'Item D', enabled: false },
-    { name: 'Item E', enabled: false },
-    { name: 'Item F', enabled: false },
-    { name: 'Item G', enabled: false },
-    { name: 'Item H', enabled: false },
-    { name: 'Item I', enabled: false },
-    { name: 'Item J', enabled: false }
+    { name: 'Item A', description: 'Add 5 seconds to the timer', category: 'powerup', price: 500, enabled: false },
+    { name: 'Item B', description: 'Reveal one random letter in the word', category: 'powerup', price: 2500, enabled: false },
+    { name: 'Item C', description: '', category: 'icon', price: 500, enabled: false },
+    { name: 'Item D', description: 'Add 15 seconds to the timer', category: 'powerup', price: 5300, enabled: false },
+    { name: 'Item E', description: '', category: 'icon', price: 500, enabled: false },
+    { name: 'Item F', description: 'Add 2 seconds to the timer', category: 'powerup', price: 500, enabled: false },
+    { name: 'Item G', description: 'Add 3 seconds to the timer', category: 'powerup', price: 100, enabled: false },
+    { name: 'Item H', description: '', category: 'icon', price: 500, enabled: false },
+    { name: 'Item I', description: 'Add 9 seconds to the timer', category: 'powerup', price: 500, enabled: false },
+    { name: 'Item J', description: '', category: 'icon', price: 2500, enabled: false }
 ];
 
 // Function to initialize the store items in the database
 const initializeStoreItems = async () => {
     try {
+        //await StoreItem.deleteMany({});
+
         // Check if store items already exist in the database
         const existingItems = await StoreItem.find();
