+++ server/routes/user.js
@@ -5,6 +5,7 @@ const UserController = require('../controllers/UserController.js');
 
 router.get('/', UserController.index)
 router.patch('/', UserController.update)
+router.patch('/inventory', UserController.createInventory)
 router.delete('/', UserController.destroy)
 
 module.exports = router
