+++ server/server.js
@@ -10,7 +10,7 @@ const words = require('./routes/words');
 const scores = require('./routes/scores');
 const AuthRoute = require('./routes/auth');
 const UserRoute = require('./routes/user');
-const InventoryRoutes = require('./routes/inventory');
+//const InventoryRoutes = require('./routes/inventory');
 const StoreRoutes = require('./routes/store');
 
 // set express app
